// src/lib/api/generated/model/notificationType.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */

/**
 * Notification type.
 */
export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  FRIEND_REQUEST: 'FRIEND_REQUEST',
  FRIEND_ACCEPT: 'FRIEND_ACCEPT',
  POST_LIKE: 'POST_LIKE',
  POST_COMMENT: 'POST_COMMENT',
  COMMENT_REPLY: 'COMMENT_REPLY',
  MENTION: 'MENTION',
  SYSTEM: 'SYSTEM',
} as const;